pyproject.toml
# Project configuration for OpenStack AI Style Guide

[project]
name = "openstack-ai-style-guide"
version = "1.0.0"
requires-python = ">=3.10"
description = "OpenStack AI Style Guide and Testing Tools"

[project.optional-dependencies]
test = [
    "stestr>=4.0.0",
    "testtools>=2.5.0",
    "fixtures>=4.0.0",
    "oslotest>=4.0.0",
    "oslo.config>=9.0.0",
    "oslo.log>=5.0.0",
    "mock>=5.0.0",
    "molecule>=6.0.0",
    "molecule-plugins>=23.0.0",
    "ansible-lint",
    "coverage>=7.0.0",
]

[tool.uv]
dev-dependencies = [
    "tox-uv>=1.0.0",
    "pre-commit",
]

# Ruff configuration for OpenStack AI Style Guide
# Configured to be compatible with OpenStack hacking rules
# See: https://docs.astral.sh/ruff/
[tool.ruff]
# Line length matching OpenStack standards (H-series rules)
line-length = 79
# Target Python 3.8+ (common OpenStack minimum)
target-version = "py38"

[tool.ruff.lint]
# Enable rule groups similar to OpenStack hacking + additional modern checks
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "D",      # pydocstyle (docstrings)
    "UP",     # pyupgrade (modern Python syntax)
    "B",      # flake8-bugbear (likely bugs and design problems)
    "C4",     # flake8-comprehensions
    "S",      # flake8-bandit (security)
    "T20",    # flake8-print (catch print statements)
    "SIM",    # flake8-simplify
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
]
# Ignore specific rules that conflict with OpenStack style or are too strict
ignore = [
    "D203",   # one-blank-line-before-class (conflicts with D211)
    "D213",   # multi-line-summary-second-line (conflicts with D212)
    "E501",   # line-too-long (handled by formatter with line-length = 79)
    "S101",   # assert-used (asserts are common in tests)
    "D100",   # Missing docstring in public module (not always needed)
    "D102",   # Missing docstring in public method (not always needed)
    "D104",   # Missing docstring in public package (not always needed)
    "D107",   # Missing docstring in __init__ (not always needed)
]

[tool.ruff.lint.per-file-ignores]
"docs/examples/bad/*.py" = ["ALL"]  # Bad examples are intentionally non-compliant
"docs/examples/good/*.py" = ["ALL"]  # Good examples may contain demonstration patterns
"tools/*.py" = ["T20"]  # Allow print statements in CLI tools
"roles/*/files/*.py" = ["T20"]  # Allow print in role scripts (CLI tools)
"tests/**/*.py" = [
    "S105",   # Possible hardcoded password (test fixtures)
    "ARG001", # Unused function argument (mocks)
]

[tool.ruff.lint.isort]
# OpenStack import ordering style
force-single-line = false
lines-after-imports = 2
known-first-party = []
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings (common in OpenStack)
convention = "google"

[tool.ruff.lint.flake8-quotes]
# OpenStack prefers single quotes for strings
inline-quotes = "single"
multiline-quotes = "single"
docstring-quotes = "double"

[tool.ruff.format]
# Formatting options compatible with OpenStack style
quote-style = "single"
indent-style = "space"
line-ending = "lf"

[tool.bandit]
# Bandit security linting configuration
exclude_dirs = [
    "tests",
    "docs/examples/bad",
    ".tox",
    ".venv",
    "venv",
]

# Skip specific tests
skips = [
    "B101",  # assert_used - common in test files
]

# Configure severity level
severity = "medium"
confidence = "medium"
